added a download option, updated the algorithm

+++ R/getECdata.R
@@ -5,9 +5,9 @@
 #' for the supplied supplied weather station IDs during the
 #' given period.
 #' @author Braeden Klaver
-#' @usage getECdata(stations, year_start, year_end, folder,
+#' @usage getECdata(stations, year_start, year_end,
 #' timeframe = c("hourly", "daily", "monthly"),
-#' verbose = TRUE, delete = TRUE)
+#' download = FALSE, folder = NULL, verbose = TRUE, delete = TRUE)
 #' @importFrom data.table fread
 #' @importFrom dplyr bind_rows
 #' @importFrom janitor clean_names
@@ -21,8 +21,9 @@
 #' @param stations Vector of weather station IDs to pull for
 #' @param year_start Starting year for data pull
 #' @param year_end End year for data pull
-#' @param folder Folder path to where you want data saved
 #' @param timeframe Timeframe of the data to pull
+#' @param download Do you want to download the data to a folder
+#' @param folder Folder path to where you want data saved
 #' @param verbose Include progress bar
 #' @param delete Delete files that failed to download or were corrupted
 #' @export getECdata
@@ -32,13 +33,18 @@
 #' https://collaboration.cmc.ec.gc.ca/cmc/climate/Get_More_Data_Plus_de_donnees/
 #' @note This function requires the other function available in the package "getECurls"
 #' @examples df <- getECdata(stations = c(52), year_start = 2022,
-#' year_end = 2023, folder = getwd(), timeframe = "daily")
+#' year_end = 2023, timeframe = "daily", download = TRUE, folder = getwd())
 
 
-getECdata <- function(stations, year_start, year_end, folder, timeframe = c("hourly", "daily", "monthly"), verbose = TRUE, delete = TRUE) {
+getECdata <- function(stations, year_start, year_end, timeframe = c("hourly", "daily", "monthly"),
+                      download = FALSE, folder = NULL, verbose = TRUE, delete = TRUE) {
+
+  if (download == TRUE & is.null(folder)) {
+    stop("Please provide a folder location to write the data to.")
+  }
 
   ## check the folder exists and try to create it if not
-  if (!file.exists(folder)) {
+  if (download == TRUE & !dir.exists(folder)) {
 
     message(paste("Creating the following folder: ", folder))
 
@@ -69,12 +75,16 @@ getECdata <- function(stations, year_start, year_end, folder, timeframe = c("hou
   sites <- unlist(lapply(urls, function(url_list) url_list$ids))
   years <- unlist(lapply(urls, function(url_list) url_list$years))
   months <- unlist(lapply(urls, function(url_list) url_list$months))
+  nfiles <- length(url_paths)
 
   ## filenames to use to save the data
-  fnames <- paste(sites, years, months, "data.csv", sep = "_")
-  fnames <- file.path(folder, fnames)
+  if (download == TRUE) {
+    fnames <- paste(sites, years, months, "data.csv", sep = "_")
+    fnames <- file.path(folder, fnames)
+  }
 
-  nfiles <- length(fnames)
+
+
 
   ## set up a progress bar if being verbose
   if (isTRUE(verbose)) {
@@ -92,68 +102,120 @@ getECdata <- function(stations, year_start, year_end, folder, timeframe = c("hou
   #iterate the download over the files
   for (i in seq_len(nfiles)) {
 
-    #get current file
-    curfile <- fnames[i]
+    if (download == TRUE) {
+      #get current file
+      curfile <- fnames[i]
 
-    #DOWNLOADING FILE
-    utils::download.file(url_paths[i], destfile = curfile, quiet = TRUE)
+      #DOWNLOADING FILE
+      utils::download.file(url_paths[i], destfile = curfile, quiet = TRUE)
 
-    ## Try reading the file
-    ecdata <- try(data.table::fread(curfile, encoding = "Latin-1", stringsAsFactors = FALSE, fill = TRUE), silent = TRUE)
-
-    if (stringr::str_detect(colnames(ecdata)[1], "DOCTYPE")) {
-
-      ecdata <- readLines(curfile, warn = FALSE) # read all lines in file
-
-      encoding <- stringi::stri_enc_detect(ecdata)$Encoding
-
-      if (is.null(encoding)) {
-
-        out$fails <- append(out$fails, url_paths[i])
+      ## Try reading the file
+      ecdata <- try(data.table::fread(curfile, encoding = "Latin-1", stringsAsFactors = FALSE, fill = TRUE), silent = TRUE)
 
 
-        if (delete) {
+      if (stringr::str_detect(colnames(ecdata)[1], "DOCTYPE")) {
 
-          file.remove(curfile) # remove file if a problem & deleting
+        ecdata <- readLines(curfile, warn = FALSE) # read all lines in file
+
+        encoding <- stringi::stri_enc_detect(ecdata)$Encoding
+
+        if (is.null(encoding)) {
+
+          out$fails <- append(out$fails, url_paths[i])
+
+
+          if (delete) {
+
+            file.remove(curfile) # remove file if a problem & deleting
+
+          }
+
+          if (isTRUE(verbose)) { # Update the progress bar
+
+            utils::setTxtProgressBar(pb, value = i)
+
+          }
+
+
+          next
 
         }
 
-        if (isTRUE(verbose)) { # Update the progress bar
+        ecdata <- iconv(ecdata, encoding, to = "UTF-8") #convert to UTF-8
+        writeLines(ecdata, curfile)          # write the data back to the file
 
-          utils::setTxtProgressBar(pb, value = i)
+        ## try to read the file again, if still an error, bail out
+        ecdata <- try(data.table::fread(curfile, encoding = "UTF-8", stringsAsFactors = FALSE), silent = TRUE)
+
+        if (inherits(ecdata, "try-error") | str_detect(colnames(ecdata)[1], "DOCTYPE")) { # yes, still!, handle read problem
+
+          out$fails <- append(out$fails, url_paths[i])    # record failed URL...
+
+
+          if (delete) {
+
+            file.remove(curfile) # remove file if a problem & deleting
+
+          }
+
+
+          if (isTRUE(verbose)) {
+
+            utils::setTxtProgressBar(pb, value = i) # update progress bar...
+
+          }
+
+          next                  # bail out of current iteration
 
         }
-
-
-        next
 
       }
 
-      ecdata <- iconv(ecdata, encoding, to = "UTF-8") #convert to UTF-8
-      writeLines(ecdata, curfile)          # write the data back to the file
+    }
 
-      ## try to read the file again, if still an error, bail out
-      ecdata <- try(data.table::fread(curfile, encoding = "UTF-8", stringsAsFactors = FALSE), silent = TRUE)
+    if(download == FALSE) {
 
-      if (inherits(ecdata, "try-error") | str_detect(colnames(ecdata)[1], "DOCTYPE")) { # yes, still!, handle read problem
+      ecdata <- try(data.table::fread(url_paths[i], encoding = "Latin-1", stringsAsFactors = FALSE, fill = TRUE), silent = TRUE)
 
-        out$fails <- append(out$fails, url_paths[i])    # record failed URL...
+      if (stringr::str_detect(colnames(ecdata)[1], "DOCTYPE")) {
+
+        ecdata <- readLines(url_paths[i], warn = FALSE) # read all lines in file
+
+        encoding <- stringi::stri_enc_detect(ecdata)$Encoding
+
+        if (is.null(encoding)) {
+
+          out$fails <- append(out$fails, url_paths[i])
 
 
-        if (delete) {
+          if (isTRUE(verbose)) { # Update the progress bar
 
-          file.remove(curfile) # remove file if a problem & deleting
+            utils::setTxtProgressBar(pb, value = i)
+
+          }
+
+
+          next
 
         }
 
+        ## try to read the file again, if still an error, bail out
+        ecdata <- try(data.table::fread(url_paths[i], encoding = "UTF-8", stringsAsFactors = FALSE, fill = TRUE), silent = TRUE)
 
-        if (isTRUE(verbose)) {
+        if (inherits(ecdata, "try-error") | str_detect(colnames(ecdata)[1], "DOCTYPE")) { # yes, still!, handle read problem
 
-          utils::setTxtProgressBar(pb, value = i) # update progress bar...
+          out$fails <- append(out$fails, url_paths[i])    # record failed URL...
+
+
+          if (isTRUE(verbose)) {
+
+            utils::setTxtProgressBar(pb, value = i) # update progress bar...
+
+          }
+
+          next                  # bail out of current iteration
 
         }
-
-        next                  # bail out of current iteration
 
       }
